test(e2e): support packaged candidate executables
