+++ tests/e2e/helpers/orca-app.ts
@@ -33,6 +33,9 @@ import {
   createElectronHomeIsolation
 } from './electron-home-isolation'
 import { createSeededTestRepo, isValidGitRepo } from './seeded-test-repo'
+import { forwardElectronProcessLogs } from './electron-process-logs'
+
+export { forwardElectronProcessLogs } from './electron-process-logs'
 
 type OrcaTestFixtures = {
   electronApp: ElectronApplication
@@ -108,27 +111,6 @@ function shouldLaunchHeadful(testInfo: TestInfo): boolean {
     return true
   }
   return testInfo.project.metadata.orcaHeadful === true
-}
-
-// Why: exported so specs that launch their own ElectronApplication outside
-// this fixture (e.g. multi-instance lifecycle tests) can still opt into the
-// same ORCA_E2E_FORWARD_APP_LOGS-gated stdout/stderr capture.
-export function forwardElectronProcessLogs(app: ElectronApplication, testInfo: TestInfo): void {
-  if (process.env.ORCA_E2E_FORWARD_APP_LOGS !== '1') {
-    return
-  }
-
-  const child = app.process()
-  const prefix = `[electron:${testInfo.title}]`
-  child.stdout?.on('data', (chunk: Buffer) => {
-    console.log(`${prefix} stdout: ${chunk.toString().trimEnd()}`)
-  })
-  child.stderr?.on('data', (chunk: Buffer) => {
-    console.error(`${prefix} stderr: ${chunk.toString().trimEnd()}`)
-  })
-  child.on('exit', (code, signal) => {
-    console.log(`${prefix} exit: code=${code ?? 'null'} signal=${signal ?? 'null'}`)
-  })
 }
 
 /**
@@ -226,8 +208,21 @@ export const test = base.extend<OrcaTestFixtures, OrcaWorkerFixtures>({
     if (recordVideoDir) {
       mkdirSync(recordVideoDir, { recursive: true })
     }
+    const packagedExecutable = process.env.ORCA_E2E_PACKAGED_EXECUTABLE?.trim()
+    const launchArgs = packagedExecutable
+      ? headful
+        ? orcaAppExtraArgs
+        : [
+            ...orcaAppExtraArgs,
+            '--disable-gpu',
+            '--disable-gpu-compositing',
+            '--disable-gpu-sandbox',
+            '--in-process-gpu'
+          ]
+      : [...orcaAppExtraArgs, ...getOrcaElectronLaunchArgs(mainPath, headful)]
     const app = await electron.launch({
-      args: [...orcaAppExtraArgs, ...getOrcaElectronLaunchArgs(mainPath, headful)],
+      ...(packagedExecutable ? { executablePath: packagedExecutable } : {}),
+      args: launchArgs,
       ...(slowMo > 0 ? { slowMo } : {}),
       ...(recordVideoDir ? { recordVideo: { dir: recordVideoDir } } : {}),
       // Why: keep NODE_ENV=development so window.__store is exposed and
